dsi: count the quantum against write data; align it to the TCP MSS
Problems: dsi_stream_receive bounded the whole DSI frame by the server
quantum, so a spec-legal FPWrite carrying exactly quantum data bytes
(dsi_len = quantum + doff) tore down the session. To stay under that
bound the tools subtracted the frame overhead from every transfer, so
clients never used the full negotiated quantum in either direction --
a throughput loss on both writes and reads from incorrect header
accounting, not a real protocol limit. dsi_writeinit copied the
buffered payload prefix into the caller's buffer (a second userspace
copy of up to the read-ahead size per write) and silently truncated
when the prefix exceeded the caller's buffer. The test client sent
each write frame as two syscalls, leaving a runt TCP segment before
every payload; transfers also ended in a runt segment whenever the
quantum was not aligned to the path MSS.

Model: per the AFP spec (DSIOpenSession, Table 1: the request quantum
is the write data size 'not including the length of the DSI header and
the AFP command'), the server quantum counts FPWrite/FPWriteExt data
only; the 16-byte DSI header and the request block ride on top. The
session quantum is resolved once per session after accept(): the
smallest 4 KiB multiple at or above the configured value whose
FPWriteExt frame (data + 36 bytes) ends in a TCP segment at least 7/8
full, growing by at most 256 KiB, capped at 256 MiB; best fill wins
when the threshold is unreachable. Buffers, inbound checks, and the
advertised value all use the grown quantum, so advertised == capacity.

Fixes: writes are bounded by dsi_len - doff with doff validated to
{12, 20}; non-write commands keep the strict bound. This restores
full-quantum transfers -- a client may again send (and request) a
complete server_quantum of data per frame, recovering the throughput
the overhead-subtraction workarounds gave up. dsi_writeinit hands the
buffered payload back by reference (NULL drains it) -- one copy, no
truncation, and afp_addicon's error path drains and returns instead of
re-arming the transfer. The client sends header, request block, and
payload in a single writev; speedtest selects FPWriteExt or legacy
FPWrite automatically from the negotiated AFP version (FPWriteExt for
AFP >= 3.0), and both tools size transfers from the full advertised
quantum with no overhead subtraction and no zero-quantum division. dsireadbuf defaults to 32, capped so dsireadbuf
x quantum stays under 1 GiB per session; server quantum values above
the maximum clamp instead of falling back to the default.
dsi_writeinit's signature change bumps the soversion to 21.

Security: the DSI frame-parsing hardening from CVE-2026-62319 and
CVE-2026-62320 (commit cc0a9d16) stays fully in place and is
strengthened -- the write data offset is now validated against an
exact {12, 20} whitelist, so malformed offsets 13-19 are rejected
instead of parsed, and dsi_tcp_open's first-frame whole-frame bound is
unchanged. Those fixes correctly reject a frame whose total exceeds
the quantum; the tools had compensated with frame-overhead
subtraction (PR #3186, PR #3187), which shrank the usable quantum.
This supersedes those workarounds by fixing the server semantics they
compensated for -- restoring full-quantum throughput -- while
preserving their invariant that the client never emits a frame the
hardened parser rejects.

Additionally closes an out-of-bounds read in the FPAddIcon write path:
the old dsi_writeinit truncated the buffered prefix into the caller's
buffer via memmove(buf, start, MIN(buflen, bytes)) but returned the
untruncated byte count, and afp_addicon then wrote that count from a
64 KiB (DSI_DATASIZ) buffer -- reading past the buffer and emitting
stale heap bytes into the icon file whenever a buffered write prefix
exceeded 64 KiB against the 1 MiB default quantum. The pointer handoff
returns the exact buffered range, removing the over-read.

Tests: afpd unit tests pin full-quantum accept (both doffs),
over-quantum reject, doff validation, the strict non-write bound, the
no-duplicate-copy handoff, the alignment search (growth, page
alignment, caps, overflow at the top of the range), and the read-ahead
clamp. RED runs for the two server bugs are preserved in the PR's CI
history. The speedtest dashboard gains sweep-average columns so
small-file regressions are visible alongside peaks.

+++ test/testsuite/speedtest_local_vfs.h
@@ -28,10 +28,10 @@ unsigned int local_createfile(CONN *conn, uint16_t vol, char type, int did,
                               char *name);
 uint16_t local_openfork(CONN *conn, uint16_t vol, int type, uint16_t bitmap,
                         int did, char *name, int access);
-unsigned int local_writeheader(DSI *dsi, uint16_t fork, int offset, int size,
-                               char *data, char whence);
-unsigned int local_writefooter(DSI *dsi, uint16_t fork, int offset, int size,
-                               char *data, char whence);
+unsigned int local_writeheader(DSI *dsi, uint16_t fork, off_t offset,
+                               size_t size, char *data, char whence);
+unsigned int local_writefooter(DSI *dsi, uint16_t fork, off_t offset,
+                               size_t size, char *data, char whence);
 unsigned int local_flushfork(CONN *conn, uint16_t fork);
 unsigned int local_closefork(CONN *conn, uint16_t fork);
 unsigned int local_delete(CONN *conn, uint16_t vol, int did, char *name);
@@ -41,10 +41,10 @@ unsigned int local_write(CONN *conn, uint16_t fork, long long offset, int size,
                          char *data, char whence);
 unsigned int local_read(CONN *conn, uint16_t fork, long long offset, int size,
                         char *data);
-unsigned int local_readheader(DSI *dsi, uint16_t fork, int offset, int size,
-                              char *data);
-unsigned int local_readfooter(DSI *dsi, uint16_t fork, int offset, int size,
-                              char *data);
+unsigned int local_readheader(DSI *dsi, uint16_t fork, off_t offset,
+                              size_t size, char *data);
+unsigned int local_readfooter(DSI *dsi, uint16_t fork, off_t offset,
+                              size_t size, char *data);
 unsigned int local_copyfile(struct CONN *conn, uint16_t svol, int sdid,
                             uint16_t dvol, int ddid, char *src, char *buf, char *dst);
 
@@ -55,16 +55,16 @@ struct vfs {
     unsigned int (*createdir)(CONN *, uint16_t, int, char *);
     unsigned int (*createfile)(CONN *, uint16_t, char, int, char *);
     uint16_t (*openfork)(CONN *, uint16_t, int, uint16_t, int, char *, int);
-    unsigned int (*writeheader)(DSI *, uint16_t, int, int, char *, char);
-    unsigned int (*writefooter)(DSI *, uint16_t, int, int, char *, char);
+    unsigned int (*writeheader)(DSI *, uint16_t, off_t, size_t, char *, char);
+    unsigned int (*writefooter)(DSI *, uint16_t, off_t, size_t, char *, char);
     unsigned int (*flushfork)(CONN *, uint16_t);
     unsigned int (*closefork)(CONN *, uint16_t);
     unsigned int (*delete)(CONN *, uint16_t, int, char *);
     unsigned int (*setforkparam)(CONN *, uint16_t,  uint16_t, off_t);
     unsigned int (*write)(CONN *, uint16_t, long long, int, char *, char);
     unsigned int (*read)(CONN *, uint16_t, long long, int, char *);
-    unsigned int (*readheader)(DSI *, uint16_t, int, int, char *);
-    unsigned int (*readfooter)(DSI *, uint16_t, int, int, char *);
+    unsigned int (*readheader)(DSI *, uint16_t, off_t, size_t, char *);
+    unsigned int (*readfooter)(DSI *, uint16_t, off_t, size_t, char *);
     unsigned int (*copyfile)(CONN *, uint16_t, int, uint16_t, int, char *, char *,
                              char *);
     uint16_t (*openvol)(CONN *, char *);
